+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Features:
   - Improved logging
   - Generate DNS Name from template for services/ingress if annotation is missing but `--fqdn-template` is specified
-  - Route 53: Support creation of records in multiple hosted zones.
+  - Route 53, Google CloudDNS: Support creation of records in multiple hosted zones.
   - Route 53: Support creation of ALIAS records when endpoint target is a ELB/ALB.
   - Ownership via TXT records
     1. Create TXT records to mark the records managed by External DNS